+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 + Adding green line around items that are clickable. (Inside the **Not Clickable Items Feature**)
 + Added **Odger waypoint** - Show the Odger waypoint when trophy fishes are in the inventory and no lava rod in hand.
 + Added the option to hide the red numbers in the scoreboard sidebar on the right side of the screen.
++ Added **Tia Relay Waypoint** - Show the next Relay waypoint for Tia The Fairy, where maintenance for the abiphone network needs to be done.
++ Added **Tia Relay Helper** - Helps with solving the sound puzzle.
 
 ## Changes
 + Hide reputation boss waypoint when damage indicator is present.